test: 跑通单元测试
