+++ packages/matman/.mocharc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  spec: 'test/scripts/**/*.ts',
+  require: 'ts-node/register',
+  reporters: true,
+  recursive: true,
+};
